Finished a basic MySQL DB Connect function.
Closes Issue #3. 
I wrote a function which uses the terminal to accept basic credentials
for connection to a MySQL database. This should help our team get
started on writing more powerful authentication methods, and gives us
something concrete to work with. The function uses the MySQL jdbc driver
to connect and authenticate. It also has a hard-coded select * function
to pull test data from our test databases inventory table. This will be
improved in the future. 

This also gave me some experience working with the Ivy and Ant xml
files, which was unchartered territory for me. I had to add an entry to
Ivy to download the mysql driver, and an entry in the ant build.xml to
show the java compiler where the mysql driver is located.