Added Postgre Support
Copied the MySQL class and re-wrote it to support PostgreSQL. Have not
been able to test the class yet, but I wanted to get it up on the
repository so that other members of the team might be able to test it
out.
Still needs to be hooked into the drop down menu on the pop-up window.