Rewrote the readEncrypt method
The original readEncrypt method was created to test the writeEncrypt
method. The original was only usable for testing because it used
variables it should not have access to. The function can now read data
from a file without nothing more than a username / alias name.

The new readEncrypt method returns an array list of bytes[]. This array
list can be used to authenticate the password and fill in information
for the user.

Closes #17