Created generic SQL execution method for MySQL.
I got some help from @rsanchez-wsu and with his advice was able to
disable FindBugs warning about variable SQL statement execution. I had
to edit the Ivy XML configuration and the Ant XML configuration to allow
the Java path to see the FindBugs jar at compile. I was then able to
turn the config() method into a true configuration method that sets up
the connection to the database, and the executeQuery() method executes a
string query and returns its results as a string.

References issue 42.