Update the executeQuery() function.
After @rsanchez-wsu suggested I update the code to use
try-with-resources, I went back through it and implemented this
functionality. There was only one connection I was not able to handle
through the try-with-resources, which is the ResultSetMetaData object.
It does not implement the AutoClose() function so it cannot be handled
by try-with-resources.

References issue #42.