issue #55 save button and issue #71 call from MainApp
Added a save and cancel button to the profile page, also moved the call
for profile to the MainApp. The save will allow us to write out new data
to the XML and the call from MainApp is the first step to allow us to
integrate better with other teams.