Made UserSettings a JavaBean #125
Because the UserSettings object represents the entire state, it makes
sense to let java automatically handle persistance. This makes
prefrences.xml the seralization of the bean. Note that the bean is harder
to edit by hand.

UserSettings.loadXmlEncodedBean and UserSettings#saveXmlEncodedBean handle
running the encoding.

Note that it is easier to add a new field than it is to change the type of
a field in a bean.