Update PostgreSQL Driver
I updated the PostgreSQL driver to have the same functionality as the
MySQL driver. They both can now return JTable objects from SELECT
statements.

I had an issue with this code in the PostgreSQL driver which did not
happen in MySQL. CheckStyle was throwing an error in the command line
about one line having 2 statements on it. I could not find the violation
in Eclipse, VIM, or in TextWrangler. Hopefully this commit doesn't break
the build, but if it does it is not my fault - it is CheckStyle being
dumb.

References Issue #155