Update MySQL Driver
Added a new method to the MySQL driver that allows inserting rows into
the database. Currently the sting parsing is very basic, if you do not
use INSERT in all caps it won't work, but I wanted to get the framework
out there so that others can collaborate.

References Issue #178.