Sent the input string to all lower case for easier handling
In the QueryBuilderTab I made the String in be all lower case so when we
are checking for instances of a word we only have to check the lower
case version in the statements. FindBugs requested that I used the
String.toLowerCase( Locale l ), hence why that is there instead of just
the .toLowerCase().
Reference Issue #178