Added multiple query execution
Implemented a regex pattern that splits an SQL script into multiple
statements in a String Array. The code then loops through the array and
executed the statements one by one. Assuming proper SQL syntax, the
queries all execute correctly.

Also changed the matching for type of queries to upper case as it is
typically the style standard for SQL.